+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * **\[core\]** Document the requirement for transactions on the methods: `ObservableValue.reportObserved()`,
   `ObservableValue.preReportChanged()`, `ObservableValue.reportChanged()`, `ComputedValue.get()` and
   `ComputedValue.reportPossiblyChanged()`.
+* **\[core\]** Ensure that when the `ReactionEnvironment` functional interface is used to wrap actions that
+  the action can not be nested. If `ArezContext.runInEnvironment(action)` is invoked when an instance of the
+  `ReactionEnvironment` functional interface is already on the call stack then the supplied procedure is directly
+  invoked without invoking the `ReactionEnvironment` functional interface again.
 
 ### [v0.108](https://github.com/arez/arez/tree/v0.108) (2018-09-27)
 [Full Changelog](https://github.com/arez/arez/compare/v0.107...v0.108)